// ACP-Checkout-Gateway/src/api.ts
export interface components {
    schemas: {
// ...
        CheckoutSession: components["schemas"]["CheckoutSessionBase"];
        CheckoutSessionWithOrder: components["schemas"]["CheckoutSessionBase"] & {
            order: components["schemas"]["Order"];
        };
        CheckoutSessionCreateRequest: {
            buyer?: components["schemas"]["Buyer"];
            items: components["schemas"]["Item"][];
            fulfillment_details?: components["schemas"]["FulfillmentDetails"];
            affiliate_attribution?: components["schemas"]["AffiliateAttribution"];
        };
        CheckoutSessionUpdateRequest: {
            buyer?: components["schemas"]["Buyer"];
            items?: components["schemas"]["Item"][];
            fulfillment_details?: components["schemas"]["FulfillmentDetails"];
            selected_fulfillment_options?: components["schemas"]["SelectedFulfillmentOption"][];
        };
        CheckoutSessionCompleteRequest: {
            buyer?: components["schemas"]["Buyer"];
            payment_data: components["schemas"]["PaymentData"];
            affiliate_attribution?: components["schemas"]["AffiliateAttribution"];
        };
// ...
    };
    responses: never;
    parameters: {
        Authorization: string;
        ContentType: string;
        AcceptLanguage: string;
        UserAgent: string;
        IdempotencyKey: string;
        RequestId: string;
        Signature: string;
        Timestamp: string;
        APIVersion: string;
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}